requirements.txt
bottle==0.12.25
cachecontrol==0.12.11
cachetools==5.0.0
certifi==2022.12.7
charset-normalizer==2.0.12
distlib==0.3.4
filelock==3.6.0
firebase-admin==5.2.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth-httplib2==0.1.0
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio-status==1.44.0
grpcio==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jaconv==0.3.4
jinja2==3.1.2
markupsafe==2.1.2; python_version >= '3.7'
msgpack==1.0.3
mypy-extensions==1.0.0; python_version >= '3.5'
mypy==1.2.0
pip==23.1.2; python_version >= '3.7'
pipenv==2022.4.30
platformdirs==2.5.2
proto-plus==1.20.3
protobuf==3.20.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==3.0.8
python-dotenv==1.0.0
requests==2.29.0
rsa==4.8
setuptools==67.7.2; python_version >= '3.7'
six==1.16.0
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.5.0; python_version >= '3.7'
uritemplate==4.1.1
urllib3==1.26.9
virtualenv-clone==0.5.7
virtualenv==20.14.1
